refactor: moving the devspace client
